[FIX] SEGV ray

+++ includes/cub3d.h
@@ -6,7 +6,7 @@
 /*   By: morishitashoto <morishitashoto@student.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/23 16:59:08 by morishitash       #+#    #+#             */
-/*   Updated: 2023/11/22 19:03:22 by morishitash      ###   ########.fr       */
+/*   Updated: 2023/11/23 06:06:02 by morishitash      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -84,7 +84,7 @@ typedef struct s_data
 	double		*length_ray; // 画面の横幅分のレイの長さ
 	int			*visual_height; // 描画する壁の高さ
 	t_direction	*wall_dir; // 壁の方向
-	int			*wall_pos; // 壁の位置
+	double		*wall_pos; // 壁の位置
 }	t_data;
 
 typedef enum e_direction
